Integrated user-via-csv & HealthCheck in Akka-Services
