+++ akka-service/src/main/resources/application.conf
@@ -1,0 +1,57 @@
+include "base-config.conf"
+
+akka.http.host = "localhost"
+akka.http.port = 8080
+akka.server.request-timeout = 60s
+akka.server.idle-timeout = 60s
+
+file.sinkDirectory = "/Users/user/Documents/shikshalokam/elevate-analytics/dev/data-pipeline/csv"
+
+postgres{
+  host = "localhost"
+  port = "5432"
+  username = "postgres"
+  password = "elevatedata"
+  database = "dev-project-analytics"
+  tables = {
+    projectsTable = ${job.env}"_projects"
+  }
+}
+
+metabase {
+  url = "http://localhost:3000/api"
+  username = "metabase-username"
+  password = "metabase-password"
+  database = "ElevateData"
+}
+
+security {
+  api-token = "8f934c7a-71b1-4ec9-98c2-d472cd9e5f1a"
+}
+
+services {
+  flink {
+    rest-api-url = "https://flink-qa.tekdinext.com"
+    jobs = [
+      "MetabaseProjectDashboardJob",
+      "ProjectsStreamJob",
+      "SurveysStreamJob",
+      "MetabaseSurveyDashboardJob",
+      "UserAndProgramManagement",
+      "ObservationsStreamJob",
+      "MetabaseObservationDashboardJob",
+      "UsersStreamJob",
+      "MentoringStreamJob",
+      "MetabaseUserDashboardJob",
+      "MetabaseMentoringDashboardJob"
+    ]
+  }
+
+  kafka {
+    broker-list = ""
+  }
+
+  metabase {
+    url = "https://qa.elevate.metabase.shikshalokam.org"
+  }
+}
