--- metabase-jobs/health-check-service/src/main/resources/application.conf
@@ -1,37 +1,0 @@
-include "base-config.conf"
-
-akka.http.host = "localhost"
-akka.http.port = 8080
-akka.server.request-timeout = 60s
-akka.server.idle-timeout = 60s
-
-services {
-  flink {
-    rest-api-url = "https://flink-qa.tekdinext.com"
-    jobs = [
-      "MetabaseProjectDashboardJob",
-      "ProjectsStreamJob",
-      "SurveysStreamJob",
-      "MetabaseSurveyDashboardJob",
-      "UserAndProgramManagement",
-      "ObservationsStreamJob",
-      "MetabaseObservationDashboardJob",
-      "UsersStreamJob",
-      "MentoringStreamJob",
-      "MetabaseUserDashboardJob",
-      "MetabaseMentoringDashboardJob"
-    ]
-  }
-
-  kafka {
-    broker-list = "localhost:9093"
-  }
-
-  metabase {
-    url = "https://qa.elevate.metabase.shikshalokam.org"
-  }
-}
-
-security {
-  api-token = ${HEALTH_CHECK_API_TOKEN}
-}
