--- metabase-jobs/health-check-service/src/main/scala/org/shikshalokam/job/healthcheck/service/models/Config.scala
@@ -1,8 +1,0 @@
-package org.shikshalokam.job.healthcheck.service.models
-
-case class FlinkConfig(restApiUrl: String, jobs: List[String])
-case class KafkaConfig(broker: String)
-case class MetabaseConfig(url: String)
-case class SecurityConfig(apiToken: String)
-
-case class AppConfig(flink: FlinkConfig, kafka: KafkaConfig, metabase: MetabaseConfig, security: SecurityConfig)
