--- metabase-jobs/health-check-service/src/main/scala/org/shikshalokam/job/healthcheck/service/models/JsonProtocol.scala
@@ -1,12 +1,0 @@
-package org.shikshalokam.job.healthcheck.service.models
-
-import spray.json._
-
-object JsonProtocol extends DefaultJsonProtocol {
-  implicit val flinkJobStatusFormat  = jsonFormat2(FlinkJobStatus)
-  implicit val flinkClusterHealthFmt = jsonFormat4(FlinkClusterHealth)
-  implicit val flinkHealthFmt        = jsonFormat2(FlinkHealth)
-  implicit val kafkaHealthFmt        = jsonFormat2(KafkaHealth)
-  implicit val metabaseHealthFmt     = jsonFormat2(MetabaseHealth)
-  implicit val fullResponseFmt       = jsonFormat4(FullHealthResponse)
-}
