+++ metabase-jobs/pom.xml
@@ -20,7 +20,6 @@
         <module>survey-dashboard-creator</module>
         <module>observation-dashboard-creator</module>
         <module>user-dashboard-creator</module>
-        <module>users-via-csv</module>
         <module>user-service</module>
         <module>mentoring-dashboard-creator</module>
     </modules>
